move extract/ tests out of tests dir for now
