+++ TESTS_TO_IMPLEMENT/extract/test_utility_data.py
@@ -2,7 +2,7 @@ import gbd_mapping as gm
 import pandas as pd
 import pytest
 
-from tests.extract.check import RUNNING_ON_CI
+from tests.conftest import RUNNING_ON_CI
 from vivarium_inputs import utility_data as ud
 
 pytestmark = pytest.mark.skipif(
