+++ tests/conftest.py
@@ -86,3 +86,16 @@ def risk_list():
 @pytest.fixture
 def locations():
     return ["Bangladesh", "Ethiopia", "Kenya", "China", "North Korea", "Nigeria"]
+
+
+def _no_gbd_access():
+    try:
+        from vivarium_inputs.globals import GbdDummy
+
+        return True
+    except ImportError:
+        pass
+    return False
+
+
+RUNNING_ON_CI = _no_gbd_access()
